backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
google-generativeai==0.3.2
langchain==0.1.0
langchain-openai==0.0.2
llama-index==0.9.8
chromadb==0.4.18

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Security and validation
python-multipart==0.0.6
email-validator==2.1.0
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Additional dependencies
websockets==12.0
jsonschema==4.21.1 